remove filter single document and rename port file
