+++ outbound/adapter/dummy-outbound-filterer.ts
@@ -1,8 +1,6 @@
 import { identity } from "fp-ts/lib/function";
-import { some } from "fp-ts/lib/Option";
-import { OutboundFilterer } from "../port/outboud-filterer";
+import { OutboundFilterer } from "../port/outbound-filterer";
 
 export const create = <I>(): OutboundFilterer<I> => ({
-  filter: some,
   filterArray: identity
 });
